Moved item tiles to the item package
