+++ SolStandard/Entity/Unit/Actions/DeployBarricadeAction.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using SolStandard.Containers.Contexts;
 using SolStandard.Entity.General;
+using SolStandard.Entity.General.Item;
 using SolStandard.Map;
 using SolStandard.Map.Elements;
 using SolStandard.Map.Elements.Cursor;
